requirements.txt
psutil>=5.9.8
scapy>=2.5.0
watchdog>=4.0.0
PyYAML>=6.0.1
requests>=2.32.3
scikit-learn>=1.5.0
pywin32>=305; platform_system == "Windows"
